+++ src/Ledger/RolloverLedger/RolloverLedger.test.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
+import user from '@testing-library/user-event';
+import { act, render, screen } from '@testing-library/react';
 import { IntlProvider } from 'react-intl';
 import { MemoryRouter } from 'react-router-dom';
 
@@ -7,6 +7,7 @@ import '@folio/stripes-acq-components/test/jest/__mock__';
 
 import { ORDER_TYPE } from '../constants';
 import RolloverLedger from './RolloverLedger';
+import { LEDGER_ROLLOVER_TYPES } from '../../common/const';
 
 const CURRENT_FISCAL_YEAR = {
   'id': '684b5dc5-92f6-4db7-b996-b549d88f5e4e',
@@ -20,31 +21,29 @@ const CURRENT_FISCAL_YEAR = {
   'series': 'FY',
 };
 
-const renderRolloverLedger = ({
-  currentFiscalYear,
-  fundTypesMap,
-  ledger,
-  onSubmit = () => { },
-  onCancel = () => { },
-  initialValues = {},
-  goToCreateFY = () => { },
-}) => (render(
+const defaultProps = {
+  onSubmit: jest.fn(),
+  onCancel: jest.fn(),
+  initialValues: {},
+  goToCreateFY: jest.fn(),
+};
+
+const renderRolloverLedger = (props = {}) => render(
   <IntlProvider locale="en">
     <MemoryRouter>
       <RolloverLedger
-        currentFiscalYear={currentFiscalYear}
-        fundTypesMap={fundTypesMap}
-        goToCreateFY={goToCreateFY}
-        initialValues={initialValues}
-        ledger={ledger}
-        onCancel={onCancel}
-        onSubmit={onSubmit}
+        {...defaultProps}
+        {...props}
       />
     </MemoryRouter>
   </IntlProvider>,
-));
+);
 
 describe('RolloverLedger', () => {
+  beforeEach(() => {
+    defaultProps.onSubmit.mockClear();
+  });
+
   it('should display form with budgets', () => {
     const initialValues = {
       budgetsRollover: [{
@@ -95,5 +94,42 @@ describe('RolloverLedger', () => {
 
     renderRolloverLedger({ currentFiscalYear: CURRENT_FISCAL_YEAR, initialValues });
     expect(screen.getByText('ui-finance.ledger.rollover.testBtn')).toBeDefined();
+  });
+
+  describe('Rollover Ledger submit', () => {
+    beforeEach(async () => {
+      const initialValues = {
+        ledgerId: '7cef8378-7cbd-1fae-bcdd-8b9d7c0af9de',
+        fromFiscalYearId: CURRENT_FISCAL_YEAR.id,
+        budgetsRollover: [],
+        encumbrancesRollover: [],
+      };
+
+      renderRolloverLedger({
+        currentFiscalYear: CURRENT_FISCAL_YEAR,
+        initialValues,
+        fiscalYears: [{
+          code: 'TY2022',
+          id: 'test-id',
+          periodStart: '2023-01-01T00:00:00.000+0000',
+        }],
+      });
+
+      const FYSelect = screen.getByRole('combobox', { name: 'ui-finance.budget.fiscalYear' });
+
+      await act(async () => user.selectOptions(FYSelect, ['test-id']));
+    });
+
+    it('should call \'handleSubmit\' for rollover when \'Rollover\' btn was clicked', async () => {
+      await act(async () => user.click(screen.getByText('ui-finance.ledger.rollover.saveBtn')));
+
+      expect(defaultProps.onSubmit).toBeCalled();
+    });
+
+    it('should call \'handleSubmit\' for rollover preview when \'Test rollover\' btn was clicked', async () => {
+      await act(async () => user.click(screen.getByText('ui-finance.ledger.rollover.testBtn')));
+
+      expect(defaultProps.onSubmit).toBeCalled();
+    });
   });
 });