+++ src/Ledger/RolloverLogs/hooks/useRolloverLogs/useRolloverLogs.test.js
@@ -39,7 +39,7 @@ describe('useRolloverLogs', () => {
       .mockReturnValue({
         get: () => ({
           json: () => Promise.resolve({
-            rolloverLogs,
+            ledgerFiscalYearRolloverLogs: rolloverLogs,
             totalRecords: rolloverLogs.length,
           }),
         }),